+++ src/lib/pages/Account.svelte
@@ -1,49 +1,34 @@
 <script>
-    import { onMount } from 'svelte';
     import Menu from '../menu/Menu.svelte';
-    import axios from '../../axiosConfig';
     import Form from '../shared/Form.svelte';
     import Input from '../shared/Input.svelte';
     import Title from '../shared/Title.svelte';
     import Link from '../shared/Link.svelte';
+    import { t } from 'svelte-i18n';
+    import { currentAccount } from '../../stores/authStore';
 
-    let username = '';
-    let email = '';
-    let role = '';
     let subscriptionEndsOn = '';
     let subscriptionCreatedOn = '';
-    let subscriptionName = '';
 
-    onMount(async () => {
-        try {
-            const { data: response } = await axios.get('/api/auth/account');
-            username = response.username;
-            email = response.email;
-            role = response.role;
-            if (response.subscription) {
-                subscriptionEndsOn = new Date(response.subscription.endAt).toLocaleDateString();
-                subscriptionCreatedOn = new Date(response.subscription.createdAt).toLocaleDateString();
-                subscriptionName = response.subscription.productName;
-            }
-        } catch (e) {
-            console.error(e);
-        }
-    });
+    $: if ($currentAccount?.subscription) {
+        subscriptionEndsOn = new Date($currentAccount.subscription.endAt).toLocaleDateString();
+        subscriptionCreatedOn = new Date($currentAccount.subscription.createdAt).toLocaleDateString();
+    }
 </script>
 
 <Menu />
 
-<Title title="Account" />
+<Title title={$t('account.title')} />
 
 <Form submittable={false}>
-    <Input label="Username" bind:value={username} disabled={true} />
-    <Input label="Email" bind:value={email} disabled={true} />
-    <Link href="/reset-password" className="text-primary-500 hover:text-white duration-300 transition-colors">Reset password</Link>
-    {#if subscriptionName && subscriptionEndsOn && subscriptionCreatedOn}
-        <Input label="Subscription" bind:value={subscriptionName} disabled={true} />
-        <Input label="Subscription ends on" bind:value={subscriptionEndsOn} disabled={true} />
-        <Input label="Subscription created on" bind:value={subscriptionCreatedOn} disabled={true} />
-    {:else if role === 'user'}
-        <Link href="/subscribe" target="_blank">Subscribe</Link>
+    <Input label={$t('common.username.label')} value={$currentAccount?.username || ''} disabled={true} />
+    <Input label={$t('common.email.label')} value={$currentAccount?.email || ''} disabled={true} />
+    {#if $currentAccount?.subscription}
+        <Input label={$t('account.subscription')} value={$currentAccount.subscription.productName} disabled={true} />
+        <Input label={$t('account.subscription-ends')} value={subscriptionEndsOn} disabled={true} />
+        <Input label={$t('account.subscribed-on')} value={subscriptionCreatedOn} disabled={true} />
+    {:else if $currentAccount?.role === 'user'}
+        <Link href="/subscribe" target="_blank">{$t('account.subscribe')}</Link>
     {/if}
+    <Link href="/reset-password" className="text-primary-500 hover:text-white duration-300 transition-colors">{$t('common.reset-password')}</Link>
 </Form>