+++ src/lib/pages/Logout.svelte
@@ -1,12 +1,15 @@
 <script>
     import { onMount } from 'svelte';
-    import { storeToast } from '../../service/toastService.js';
+    import { showToast } from '../../service/toastService.js';
+    import { navigate } from 'svelte-routing';
+    import { clearCurrentAccount } from '../../stores/authStore.js';
 
     onMount(() => {
         document.title = 'Logout';
         localStorage.removeItem('apiToken');
         localStorage.removeItem('subscribed');
-        storeToast('Logged out', 'success');
-        window.location.href = '/';
+        clearCurrentAccount();
+        showToast('Logged out', 'success');
+        navigate('/login');
     });
 </script>