+++ src/lib/pages/Subscribe.svelte
@@ -2,8 +2,9 @@
     import Menu from '../menu/Menu.svelte';
     import Title from '../shared/Title.svelte';
     import Subtitle from '../shared/Subtitle.svelte';
+    import { t } from 'svelte-i18n';
 </script>
 
 <Menu />
-<Title title="Subscribe" />
-<Subtitle>Coming soon ...</Subtitle>
+<Title title={$t('subscribe.title')} />
+<Subtitle>{$t('common.coming-soon')}</Subtitle>