+++ src/lib/tassadecks/EditorCardDetails.svelte
@@ -4,6 +4,7 @@
     import Select from '../shared/Select.svelte';
     import Button from '../shared/Button.svelte';
     import CardBadge from './CardBadge.svelte';
+    import { t } from 'svelte-i18n';
 
     const dispatch = createEventDispatcher();
 
@@ -65,7 +66,7 @@
                     </div>
                 {/if}
                 {#if !isLegal}
-                    <CardBadge type="error">Banned</CardBadge>
+                    <CardBadge type="error">{$t('badge.banned')}</CardBadge>
                 {/if}
             </div>
         </div>
@@ -85,12 +86,16 @@
                     <IconButton icon="plus" size={32} disabled={!isBasicLand} on:click={() => dispatch('cardIncrement', selectedCard)} />
                 </div>
                 <div class="flex justify-center">
-                    <Button ariaLabel="Switch print" on:click={() => (switching = true)}>Switch print</Button>
+                    <Button ariaLabel={$t('tassadecks.editor.print.switch')} on:click={() => (switching = true)}
+                        >{$t('tassadecks.editor.print.switch')}</Button
+                    >
                 </div>
             {:else}
                 <div class="flex justify-center gap-10">
                     <IconButton icon="trash" size={32} on:click={() => dispatch('cardDecrement', selectedCard)} />
-                    <Button ariaLabel="Switch print" on:click={() => (switching = true)}>Switch print</Button>
+                    <Button ariaLabel={$t('tassadecks.editor.print.switch')} on:click={() => (switching = true)}
+                        >{$t('tassadecks.editor.print.switch')}</Button
+                    >
                 </div>
             {/if}
         </div>