+++ src/lib/tassadecks/NewDeck.svelte
@@ -2,9 +2,10 @@
     import Title from '../shared/Title.svelte';
     import Menu from '../menu/Menu.svelte';
     import DeckForm from './DeckForm.svelte';
+    import { t } from 'svelte-i18n';
 </script>
 
 <Menu />
-<Title title="New deck" />
+<Title title={$t('tassadecks.new.title')} />
 
 <DeckForm scratch={false} />